+++ include/light/sphereLightSource.h
@@ -8,17 +8,17 @@ class SphereLightSource : public LightSource {
 public:
     SphereLightSource()= default;
 
-    SphereLightSource(double r, Point3f p, Vec3f emission, Vec3f kd = Vec3f(), Vec3f ks = Vec3f())
+    SphereLightSource(Float r, Point3f p, Vec3f emission, Vec3f kd = Vec3f(), Vec3f ks = Vec3f())
             : radius(r), position(p), LightSource(emission, kd, ks) {}
 
     ~SphereLightSource() override = default;
 
-    double intersect(const Ray &ray) const override;
+    Float intersect(const Ray &ray) const override;
 
     void clear() override;
 
     Point3f position;
-    double radius;
+    Float radius;
 };
 
 #endif //LIGHT_SPHERELIGHTSOURCE_H
